+++ docs/changelog.rst
@@ -5,9 +5,9 @@ v0.6.0 (in development)
 -----------------------
 - Require ``javaproperties`` 0.6.\*
 - Use ``'javapropertiesreplace'`` error handler when opening output files
-- Gave :program:`javaproperties`' :program:`format` and :program:`select`
-  subcommands ``--ascii`` and ``--unicode`` options for controlling the
-  escaping of non-ASCII characters in output
+- Gave :program:`javaproperties`' :program:`format`, :program:`select`, and
+  :program:`set` subcommands ``--ascii`` and ``--unicode`` options for
+  controlling the escaping of non-ASCII characters in output
 - Gave :program:`properties2json` ``--ascii`` and ``--unicode`` options for
   controlling the escaping of non-ASCII characters in output
 - :program:`properties2json` now always outputs UTF-8